moves wallet rpc util functions from types to utils
we have a couple of functions defined in the 'types' module. functions like
these may fit better in 'utils' than in 'types'
